refactor: implement multiple small review suggestions
