+++ src/ElectronBackend/main/menu.ts
@@ -5,9 +5,8 @@
 // SPDX-License-Identifier: Apache-2.0
 import { app, BrowserWindow, Menu, shell } from 'electron';
 
-import { importFileFormats } from '../../Frontend/shared-constants';
 import { AllowedFrontendChannels } from '../../shared/ipc-channels';
-import { ExportType } from '../../shared/shared-types';
+import { ExportType, FileFormatInfo } from '../../shared/shared-types';
 import { isFileLoaded } from '../utils/getLoadedFile';
 import { getGlobalBackendState } from './globalBackendState';
 import {
@@ -81,6 +80,10 @@ const INITIALLY_DISABLED_ITEMS_INFO: Record<
     id: 'search-resources-linked',
   },
 };
+
+export const importFileFormats: Array<FileFormatInfo> = [
+  ['Legacy Opossum File', ['json', 'json.gz']],
+];
 
 export async function createMenu(mainWindow: BrowserWindow): Promise<Menu> {
   const webContents = mainWindow.webContents;
